Separate TestUnpywallCache and TestUnpywall classes into different files
